chore: split E2E tests into per-subsystem files, update package comment
Split the 1,514-line e2e/lifecycle_test.go monolith into 7 focused files:
- helpers_test.go: testStack, newTestStack, poll helpers
- registration_test.go: scenarios 1-3 (registration, assignment)
- lifecycle_test.go: scenarios 4-5 (full lifecycle, heartbeat)
- dag_test.go: scenarios 6-7 (linear DAG, fork-join DAG)
- crash_policy_test.go: scenarios 8-12 (backoff, circuit breaker, cooldown)
- error_paths_test.go: scenarios 13-22 (error injection, re-enqueue)
- stress_test.go: scenario 23 (concurrent crash+complete)

Updated package comment to name all three export_e2e.go shim functions
(ApplyEventForTest, CrashAgentForTest, CompleteAgentForTest) and clarify
precondition-setup vs scenario-bypass usage patterns.

Added [gRPC-bypassed] annotation to stress test doc comment.

Council round 2 recommendation file included.
